+++ spec/guard/dsl_spec.rb
@@ -325,7 +325,7 @@ describe Guard::Dsl do
     disable_user_config
 
     it 'adds the paths to the listener\'s ignore_paths' do
-      ::Guard::UI.should_receive(:deprecation).with(described_class::IGNORE_PATHS_DEPRECATION)
+      ::Guard::UI.should_receive(:deprecation).with(Guard::Deprecator::DSL_METHOD_IGNORE_PATHS_DEPRECATION)
 
       described_class.evaluate_guardfile(:guardfile_contents => 'ignore_paths \'foo\', \'bar\'')
     end
@@ -402,18 +402,18 @@ describe Guard::Dsl do
     disable_user_config
 
     it 'disables the interactions with :off' do
-      ::Guard::UI.should_not_receive(:deprecation).with(described_class::INTERACTOR_DEPRECATION)
+      ::Guard::UI.should_not_receive(:deprecation).with(Guard::Deprecator::DSL_METHOD_INTERACTOR_DEPRECATION)
       described_class.evaluate_guardfile(:guardfile_contents => 'interactor :off')
       Guard::Interactor.enabled.should be_false
     end
 
     it 'shows a deprecation for symbols other than :off' do
-      ::Guard::UI.should_receive(:deprecation).with(described_class::INTERACTOR_DEPRECATION)
+      ::Guard::UI.should_receive(:deprecation).with(Guard::Deprecator::DSL_METHOD_INTERACTOR_DEPRECATION)
       described_class.evaluate_guardfile(:guardfile_contents => 'interactor :coolline')
     end
 
     it 'passes the options to the interactor' do
-      ::Guard::UI.should_not_receive(:deprecation).with(described_class::INTERACTOR_DEPRECATION)
+      ::Guard::UI.should_not_receive(:deprecation).with(Guard::Deprecator::DSL_METHOD_INTERACTOR_DEPRECATION)
       described_class.evaluate_guardfile(:guardfile_contents => 'interactor :option1 => \'a\', :option2 => 123')
       Guard::Interactor.options.should include({ :option1 => 'a', :option2 => 123 })
     end